Move parameter factory into traject from converter.
