+++ morepath/traject.py
@@ -30,8 +30,9 @@ For a description of a similar algorithm also read: http://littledev.nl/?p=99
 import re
 from functools import total_ordering
 from reg import arginfo
+from webob.exc import HTTPBadRequest
 
-from .converter import IDENTITY_CONVERTER, ParameterFactory
+from .converter import IDENTITY_CONVERTER
 from .error import TrajectError
 from .mapply import mapply
 
@@ -395,6 +396,65 @@ class TrajectRegistry(object):
         result = mapply(node.model_factory, **variables)
         if result is None:
             return None
+        return result
+
+
+class ParameterFactory(object):
+    """Convert URL parameters.
+
+    Given expected URL parameters, converters for them and required
+    parameters, create a dictionary of converted URL parameters with
+    Python values.
+
+    :param parameters: dictionary of parameter names -> default values.
+    :param converters: dictionary of parameter names -> converters.
+    :param required: dictionary of parameter names -> required booleans.
+    :param extra: should extra unknown parameters be included?
+    """
+    def __init__(self, parameters, converters, required, extra=False):
+        self.parameters = parameters
+        self.converters = converters
+        self.required = required
+        self.extra = extra
+
+    def __call__(self, request):
+        """Convert URL parameters to Python dictionary with values.
+        """
+        result = {}
+        url_parameters = request.GET
+        for name, default in self.parameters.items():
+            value = url_parameters.getall(name)
+            converter = self.converters.get(name, IDENTITY_CONVERTER)
+            if converter.is_missing(value):
+                if name in self.required:
+                    raise HTTPBadRequest(
+                        "Required URL parameter missing: %s" %
+                        name)
+                result[name] = default
+                continue
+            try:
+                result[name] = converter.decode(value)
+            except ValueError:
+                raise HTTPBadRequest(
+                    "Cannot decode URL parameter %s: %s" % (
+                        name, value))
+
+        if not self.extra:
+            return result
+
+        remaining = set(url_parameters.keys()).difference(
+            set(result.keys()))
+        extra = {}
+        for name in remaining:
+            value = url_parameters.getall(name)
+            converter = self.converters.get(name, IDENTITY_CONVERTER)
+            try:
+                extra[name] = converter.decode(value)
+            except ValueError:
+                raise HTTPBadRequest(
+                    "Cannot decode URL parameter %s: %s" % (
+                        name, value))
+        result['extra_parameters'] = extra
         return result
